create a function to run terraform

+++ manage.sh
@@ -1,29 +1,38 @@
 #!/bin/bash
 if [[ "$#" -lt 1 ]]; then
-  echo "usage: $(basename "$0") istio|linkerd [destroy|shutdown|stop|start|info]"
+  echo "usage: $(basename "$0") istio|linkerd [create|destroy|stop|start|list]"
   exit 1
 fi
 
 SM=$1
-ACTION=${2:-apply}
+ACTION=$2
+
+run_terraform() {
+  if ! [[ "$1" =~ apply|destroy ]]; then
+    echo "Error: run_terraform(): $1 is not valid input"
+    exit 1
+  fi
+  rm -rf .terraform
+  terraform -chdir=terraform init -backend-config=../"$2"/backend.tfvars
+  terraform -chdir=terraform "$1" -var-file=../"$2"/terraform.tfvars
+}
 
 case "$ACTION" in
+  create)
+    run_terraform 'apply' "$SM"
+    ;;
+  destroy)
+    run_terraform 'destroy' "$SM"
+    ;;
   stop)
     aws lightsail get-instances --query 'instances[*].name' --output json | jq .[] | grep istio | xargs -n1 aws lightsail stop-instance --instance-name
-    exit
     ;;
   start)
     aws lightsail get-instances --query 'instances[*].name' --output json | jq .[] | grep istio | xargs -n1 aws lightsail start-instance --instance-name
     aws lightsail get-instances --query 'instances[*].[publicIpAddress,name]'
-    exit
     ;;
-  info)
+  list)
     aws lightsail get-instances --query 'instances[*].[publicIpAddress,name]'
-    exit
     ;;
   *)
 esac
-
-rm -rf .terraform
-terraform -chdir=terraform init -backend-config=../"$SM"/backend.tfvars
-terraform -chdir=terraform "$ACTION" -var-file=../"$SM"/terraform.tfvars
